workspace/apt_requirements.txt
python3-serial
python3-rpi-lgpio